+++ lib/constants/routes.dart
@@ -2,3 +2,4 @@ const loginRoute = '/login/';
 const registerRoute = '/register/';
 const notesRoute = '/notes/';
 const verifyEmailRoute = '/verifyemail/';
+const newNoteRoute = '/notes/new-note/';
