+++ lib/views/notes/new_notes_view.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+class NewNoteView extends StatefulWidget {
+  const NewNoteView({Key? key}) : super(key: key);
+
+  @override
+  State<NewNoteView> createState() => _NewNoteViewState();
+}
+
+class _NewNoteViewState extends State<NewNoteView> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text("New note")),
+      body: const Text("Write your notes here"),
+    );
+  }
+}
