--- lib/views/notes_view.dart
@@ -1,80 +1,0 @@
-// ignore_for_file: use_build_context_synchronously
-import 'package:flutter/material.dart';
-import 'package:notes/main.dart';
-import 'package:notes/services/crud/notes_service.dart';
-import '../constants/routes.dart';
-import '../enums/menu_actions.dart';
-import '../services/auth/auth_service.dart';
-
-class NotesView extends StatefulWidget {
-  final String name = "Gaurav";
-  const NotesView({Key? key}) : super(key: key);
-
-  @override
-  State<NotesView> createState() => _NotesViewState();
-}
-
-class _NotesViewState extends State<NotesView> {
-  late final NotesService _notesService;
-  String get userEmail => AuthService.firebase().currentUser!.email!;
-  @override
-  void initState() {
-    _notesService = NotesService();
-    _notesService.open();
-    super.initState();
-  }
-
-  @override
-  void dispose() {
-    _notesService.close();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: AppBar(
-          title: const Text("Notes"),
-          actions: [
-            PopupMenuButton(onSelected: (value) async {
-              switch (value) {
-                case MenuAction.logout:
-                  final logout = await logoutDialog(context);
-                  if (logout) {
-                    await AuthService.firebase().logOut();
-                    Navigator.of(context)
-                        .pushNamedAndRemoveUntil(loginRoute, (route) => false);
-                  }
-              }
-            }, itemBuilder: (context) {
-              return const [
-                PopupMenuItem<MenuAction>(
-                  value: MenuAction.logout,
-                  child: Text("Log out"),
-                )
-              ];
-            })
-          ],
-        ),
-        body: FutureBuilder(
-            future: _notesService.getOrCreateUser(email: userEmail),
-            builder: (context, snapshot) {
-              switch (snapshot.connectionState) {
-                case ConnectionState.done:
-                  return StreamBuilder(
-                    stream: _notesService.allNotes,
-                    builder: (context, snapshot) {
-                      switch (snapshot.connectionState) {
-                        case ConnectionState.waiting:
-                          return const Text("Waiting for all notes");
-                        default:
-                          return const CircularProgressIndicator();
-                      }
-                    },
-                  );
-                default:
-                  return const CircularProgressIndicator();
-              }
-            }));
-  }
-}
