+++ test/auth_test.dart
@@ -98,7 +98,7 @@ class MockAuthProvider implements AuthProvider {
     if (email == 'aaa@aaa.aaa') throw UserNotFoundAuthException();
     if (password == 'simplepass') throw WrongPasswordAuthException();
     if (!isInitialized) throw NotInitializedException();
-    const user = AuthUser(isEmailVerified: false);
+    const user = AuthUser(isEmailVerified: false, email: '');
     _user = user;
     return Future.value(user);
   }
@@ -116,7 +116,8 @@ class MockAuthProvider implements AuthProvider {
     if (!isInitialized) throw NotInitializedException();
     final user = _user;
     if (user == null) throw UserNotFoundAuthException();
-    const newUser = AuthUser(isEmailVerified: true);
+    const newUser =
+        AuthUser(isEmailVerified: true, email: 'gauravpatil8604@gmail.com');
     _user = newUser;
   }
 }
